nym-connect: extract out events.rs
