+++ nym-connect/src-tauri/src/tasks.rs
@@ -14,9 +14,10 @@ use nym_socks5::client::{config::Config as Socks5Config, Socks5ControlMessageSen
 
 use crate::{
     error::Result,
+    events::{self, emit_event, emit_status_event},
     models::{ConnectionStatusKind, ConnectivityTestResult},
     operations,
-    state::{GatewayConnectivity, State},
+    state::State,
 };
 
 pub type ExitStatusReceiver = futures::channel::oneshot::Receiver<Socks5ExitStatusMessage>;
@@ -91,30 +92,6 @@ pub fn start_nym_socks5_client(
     ))
 }
 
-#[derive(Clone, serde::Serialize)]
-struct Payload {
-    title: String,
-    message: String,
-}
-
-impl Payload {
-    fn new(title: String, message: String) -> Self {
-        Self { title, message }
-    }
-}
-
-fn emit_event(event: &str, title: &str, msg: &str, window: &tauri::Window<tauri::Wry>) {
-    if let Err(err) = window.emit(event, Payload::new(title.into(), msg.into())) {
-        log::error!("Failed to emit tauri event: {err}");
-    }
-}
-
-fn emit_status_event<T: ToString>(event: &str, msg: &T, window: &tauri::Window<tauri::Wry>) {
-    if let Err(err) = window.emit(event, Payload::new("SOCKS5 update".into(), msg.to_string())) {
-        log::error!("Failed to emit tauri event: {err}");
-    }
-}
-
 pub fn start_connection_check(state: Arc<RwLock<State>>, window: tauri::Window<tauri::Wry>) {
     log::debug!("Starting connection check handler");
     tokio::spawn(async move {
@@ -155,43 +132,6 @@ pub fn start_connection_check(state: Arc<RwLock<State>>, window: tauri::Window<t
     });
 }
 
-async fn handle_task_status(
-    task_status: &TaskStatus,
-    state: &Arc<RwLock<State>>,
-    window: &tauri::Window,
-) {
-    match task_status {
-        TaskStatus::Ready => {
-            {
-                let mut state_w = state.write().await;
-                state_w.mark_connected(window);
-            }
-
-            emit_status_event("socks5-connected-event", task_status, window);
-            start_connection_check(state.clone(), window.clone());
-        }
-    }
-}
-
-async fn handle_client_status_message(
-    client_status_message: &ClientCoreStatusMessage,
-    state: &Arc<RwLock<State>>,
-    window: &tauri::Window,
-) {
-    // TODO: use this instead once we change on the frontend too
-    let _event_name = match client_status_message {
-        ClientCoreStatusMessage::GatewayIsSlow | ClientCoreStatusMessage::GatewayIsVerySlow => {
-            "socks5-gateway-status"
-        }
-    };
-
-    if let Ok(connectivity) = GatewayConnectivity::try_from(client_status_message) {
-        state.write().await.set_gateway_connectivity(connectivity);
-    }
-
-    emit_status_event("socks5-status-event", client_status_message, window);
-}
-
 /// The status listener listens for non-exit status messages from the background socks5 proxy task.
 pub fn start_status_listener(
     state: Arc<RwLock<State>>,
@@ -205,11 +145,11 @@ pub fn start_status_listener(
             log::info!("SOCKS5 proxy sent status message: {}", msg);
 
             if let Some(task_status) = msg.downcast_ref::<TaskStatus>() {
-                handle_task_status(task_status, &state, &window).await;
+                events::handle_task_status(task_status, &state, &window).await;
             } else if let Some(client_status_message) =
                 msg.downcast_ref::<ClientCoreStatusMessage>()
             {
-                handle_client_status_message(client_status_message, &state, &window).await;
+                events::handle_client_status_message(client_status_message, &state, &window).await;
             } else {
                 emit_status_event("socks5-status-event", &msg, &window);
             }
